+++ lib/ZarrCore/src/metadata.jl
@@ -292,6 +292,7 @@ fill_value_decoding(v, T) = T(v)
 fill_value_decoding(v::Number, T::Type{String}) = v == 0 ? "" : T(UInt8[v])
 fill_value_decoding(v, ::Type{ASCIIChar}) = v == "" ? nothing : v
 fill_value_decoding(v::Nothing, ::Type{ZarrCore.ASCIIChar}) = v
+fill_value_decoding(v::Vector, T::Type{<:Complex}) = T(v[1], v[2])
 # Sometimes when translating between CF (climate and forecast) convention data
 # and Zarr groups, fill values are left as "negative integers" to encode unsigned
 # integers.  So, we have to convert to the signed type with the same number of bytes